# sysorder/src/main/java/com/br/refera/sysorder/model/ErrorResponse.java
package com.br.refera.sysorder.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ErrorResponse {
    private String code;
    private String description;
    private String message;
}